fix: share & log
