+++ route/handler.go
@@ -4,59 +4,14 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"net/http"
 	"net/url"
 
 	"github.com/gin-gonic/gin"
 	"github.com/gorilla/websocket"
 	log "github.com/sirupsen/logrus"
-	"github.com/yudai/gotty/webtty"
 
 	"github.com/wrfly/container-web-tty/types"
-	"github.com/wrfly/container-web-tty/util"
 )
-
-func (server *Server) handleExecRedirect(c *gin.Context) {
-	containerID := c.Param("id")
-	execID := server.setContainerID(containerID)
-	c.Redirect(302, "/exec/"+execID)
-}
-
-func (server *Server) handleWSIndex(c *gin.Context) {
-	execID := c.Param("id")
-	containerID, ok := server.getContainerID(execID)
-	if !ok {
-		log.Errorf("exec id %s not found", execID)
-		return
-	}
-	cInfo := server.containerCli.GetInfo(c.Request.Context(), containerID)
-	titleVars := server.titleVariables(
-		[]string{"server"},
-		map[string]map[string]interface{}{
-			"server": map[string]interface{}{
-				"containerName": cInfo.Name,
-			},
-		},
-	)
-
-	titleBuf := new(bytes.Buffer)
-	err := titleTemplate.Execute(titleBuf, titleVars)
-	if err != nil {
-		c.Error(err)
-	}
-
-	indexVars := map[string]interface{}{
-		"title": titleBuf.String(),
-	}
-
-	indexBuf := new(bytes.Buffer)
-	err = indexTemplate.Execute(indexBuf, indexVars)
-	if err != nil {
-		c.Error(err)
-	}
-
-	c.Writer.Write(indexBuf.Bytes())
-}
 
 func (server *Server) handleAuthToken(c *gin.Context) {
 	c.Header("Content-Type", "application/javascript")
@@ -147,79 +102,6 @@ func (server *Server) handleStopContainer(c *gin.Context) {
 func (server *Server) handleRestartContainer(c *gin.Context) {
 	server.handleContainerActions(c, "restart")
 }
-
-func (server *Server) handleLogs(c *gin.Context) {
-	ctx := c.Request.Context()
-
-	conn, err := server.upgrader.Upgrade(c.Writer, c.Request, nil)
-	if err != nil {
-		c.String(http.StatusInternalServerError, "server error: %s", err)
-		return
-	}
-	defer conn.Close()
-
-	initArg, err := server.readInitMessage(conn)
-	if err != nil {
-		c.String(http.StatusBadRequest, "read init message error: %s", err)
-		return
-	}
-
-	q, err := parseQuery(initArg)
-	if err != nil {
-		c.String(http.StatusBadRequest, err.Error())
-		return
-	}
-	follow := true
-	if v := q.Get("follow"); v != "1" && v != "" {
-		follow = false
-	}
-	tail := "10"
-	if v := q.Get("tail"); v != "" {
-		tail = v
-	}
-	opts := types.LogOptions{
-		ID:     c.Param("id"),
-		Follow: follow,
-		Tail:   tail,
-	}
-
-	container := server.containerCli.GetInfo(ctx, opts.ID)
-
-	log.Debugf("get logs of container: %s", container.ID)
-	logsReadCloser, err := server.containerCli.Logs(ctx, opts)
-	if err != nil {
-		c.String(http.StatusInternalServerError, "get logs error: %s", err)
-		return
-	}
-	defer logsReadCloser.Close()
-
-	titleBuf, err := server.makeTitleBuff(container)
-	if err != nil {
-		c.String(http.StatusInternalServerError, "failed to fill window title template: %s", err)
-		return
-	}
-
-	tty, err := webtty.New(
-		&wsWrapper{conn},
-		newSlave(util.NopRWCloser(logsReadCloser), false),
-		[]webtty.Option{
-			webtty.WithWindowTitle(titleBuf),
-			webtty.WithPermitWrite(), // can type "enter"
-		}...,
-	)
-	if err != nil {
-		c.String(http.StatusInternalServerError, "failed to create webtty: %s", err)
-		return
-	}
-
-	if err := tty.Run(ctx); err != nil {
-		if err != webtty.ErrMasterClosed && err != webtty.ErrSlaveClosed {
-			log.Errorf("failed to run webtty: %s", err)
-		}
-	}
-}
-
-func (server *Server) terminalPage(c *gin.Context) { server.handleWSIndex(c) }
 
 func (server *Server) makeTitleBuff(c types.Container) ([]byte, error) {
 	location := "127.0.0.1"
